--- pyapi/endpoints/users.py
@@ -1,7 +1,0 @@
-from fastapi import APIRouter
-
-router = APIRouter()
-
-@router.get("/users")
-async def get_users():
-    return {"users": ["User 1", "User 2"]}
